keepalived_status_check.py からロジック部分を status_check.py に切り出した
