+++ elements/keepalived-status-check/pre-install.d/10-keepalived-status-check-install
@@ -7,7 +7,8 @@ set -eu
 set -o pipefail
 
 install -m 0755 -o root -g root $(dirname $0)/keepalived_status_check.py /usr/local/bin/keepalived_status_check.py
+install -m 0755 -o root -g root $(dirname $0)/status_check.py /usr/local/bin/status_check.py
 sed -i -e 's,%%{OS_USERNAME}%%,'${OS_USERNAME}',g' \
        -e 's,%%{OS_PASSWORD}%%,'${OS_PASSWORD}',g' \
        -e 's,%%{OS_AUTH_URL}%%,'${OS_AUTH_URL}',g' \
-       /usr/local/bin/keepalived_status_check.py
+       /usr/local/bin/status_check.py
